fix: incidents for inactive/parked assets
No longer creates incidents for inactive/parked assets.

AssetMock didn't send back a proper response, but the problem wasn't
visible in the tests (except for stacktraces) due to how the error
handling in AssetCommunicationBean is set up. AssetClient already has a
method for creating a poll, but exception handling and the username are
handled differently. Using the AssetClient method would be better as
mocking becomes easier and cleaner, but would have to be changed outside
of this bug fix.

Refs: FART-534